requirements.txt
fastapi
pydantic
requests
beautifulsoup4
streamlit
python-dotenv
sumy
nltk
google-generativeai

